fix: extract workflow save prompt to composable
