+++ packages/editor-ui/src/views/NodeViewSwitcher.vue
@@ -5,31 +5,17 @@ import { onBeforeRouteLeave, useRoute, useRouter } from 'vue-router';
 import NodeViewV1 from '@/views/NodeView.vue';
 import NodeViewV2 from '@/views/NodeView.v2.vue';
 import { getNodeViewTab } from '@/utils/canvasUtils';
-import {
-	MAIN_HEADER_TABS,
-	MODAL_CANCEL,
-	MODAL_CONFIRM,
-	PLACEHOLDER_EMPTY_WORKFLOW_ID,
-	VIEWS,
-} from '@/constants';
-import { useUIStore } from '@/stores/ui.store';
-import { useMessage } from '@/composables/useMessage';
-import { useI18n } from '@/composables/useI18n';
+import { MAIN_HEADER_TABS, PLACEHOLDER_EMPTY_WORKFLOW_ID, VIEWS } from '@/constants';
 import { useWorkflowsStore } from '@/stores/workflows.store';
-import { useNpsSurveyStore } from '@/stores/npsSurvey.store';
 import { useWorkflowHelpers } from '@/composables/useWorkflowHelpers';
 import { useCanvasOperations } from '@/composables/useCanvasOperations';
 import { useSourceControlStore } from '@/stores/sourceControl.store';
 
-const uiStore = useUIStore();
 const workflowsStore = useWorkflowsStore();
-const npsSurveyStore = useNpsSurveyStore();
 const sourceControlStore = useSourceControlStore();
 
 const router = useRouter();
 const route = useRoute();
-const message = useMessage();
-const i18n = useI18n();
 const workflowHelpers = useWorkflowHelpers({ router });
 
 const { resetWorkspace } = useCanvasOperations({ router });
@@ -56,32 +42,17 @@ onBeforeRouteLeave(async (to, from, next) => {
 	if (
 		toNodeViewTab === MAIN_HEADER_TABS.EXECUTIONS ||
 		from.name === VIEWS.TEMPLATE_IMPORT ||
-		(toNodeViewTab === MAIN_HEADER_TABS.WORKFLOW && from.name === VIEWS.EXECUTION_DEBUG)
+		(toNodeViewTab === MAIN_HEADER_TABS.WORKFLOW && from.name === VIEWS.EXECUTION_DEBUG) ||
+		isReadOnlyEnvironment.value
 	) {
 		next();
 		return;
 	}
 
-	if (uiStore.stateIsDirty && !isReadOnlyEnvironment.value) {
-		const confirmModal = await message.confirm(
-			i18n.baseText('generic.unsavedWork.confirmMessage.message'),
-			{
-				title: i18n.baseText('generic.unsavedWork.confirmMessage.headline'),
-				type: 'warning',
-				confirmButtonText: i18n.baseText('generic.unsavedWork.confirmMessage.confirmButtonText'),
-				cancelButtonText: i18n.baseText('generic.unsavedWork.confirmMessage.cancelButtonText'),
-				showClose: true,
-			},
-		);
-
-		if (confirmModal === MODAL_CONFIRM) {
+	await workflowHelpers.promptSaveUnsavedWorkflowChanges(next, {
+		async confirm() {
 			// Make sure workflow id is empty when leaving the editor
 			workflowsStore.setWorkflowId(PLACEHOLDER_EMPTY_WORKFLOW_ID);
-			const saved = await workflowHelpers.saveCurrentWorkflow({}, false);
-			if (saved) {
-				await npsSurveyStore.fetchPromptsData();
-			}
-			uiStore.stateIsDirty = false;
 
 			if (from.name === VIEWS.NEW_WORKFLOW) {
 				// Replace the current route with the new workflow route
@@ -92,18 +63,17 @@ onBeforeRouteLeave(async (to, from, next) => {
 				});
 
 				await router.push(to);
-			} else {
-				next();
+
+				return false;
 			}
-		} else if (confirmModal === MODAL_CANCEL) {
+
+			return true;
+		},
+		async cancel() {
 			workflowsStore.setWorkflowId(PLACEHOLDER_EMPTY_WORKFLOW_ID);
 			resetWorkspace();
-			uiStore.stateIsDirty = false;
-			next();
-		}
-	} else {
-		next();
-	}
+		},
+	});
 });
 </script>
